Getting rid of the notion of action.
With whitelist and blacklist, there really isn't much of a reason to
have the action around anymore. It is just an extra conditional that
doesn't make sense. I'm removing it now in the nginx module and in the
test suite. I will remove it from the Apache bits after some additional testing.

Signed-off-by: Aaron Bedra <aaron@aaronbedra.com>